Replace type of "dp_mem_alloc" from 16 bits to 32 bits. Indeed, we may have huge ADN sequences (not the case for proteins) with potential huge alignments, bigger than 2^16 nucleotides.
This can lead to crash since the "score_array" may be too small.

With 32 bits, we should be OK.

+++ src/alignment/tools/impl/SemiGappedAlign.cpp
@@ -129,8 +129,8 @@ int SemiGapAlign::compute (
     ScoreInt next_score    = 0;
     ScoreInt best_score    = score;
 
-    u_int16_t num_extra_cells;
-    u_int16_t dp_mem_alloc = 1000;
+    u_int32_t num_extra_cells;
+    u_int32_t dp_mem_alloc = 1000;
 
     /* do initialization and sanity-checking */
     *a_offset = 0;
